Add SearchBar component and update color scheme
A new SearchBar component is introduced in the Header to improve navigation. Additionally, a slight update is made to the color scheme in the tailwind.config.js file where a "lightgrey" color is added and the value of "grey" is adjusted. New files for "m-interaction" and "m-chat-probe" are also created in the docs.